Using NuGet.ProjectModel from the install path to avoid version conflicts (#566)
This change removes the dependency on NuGet.ProjectModel from the LanguageServer project and instead relies on the installed version of the binary. This change intends to avoid the dependency between the .Net Core SDK version and NuGet in MSBuildLocator. See issue #86 for additional reference.
(https://github.com/microsoft/MSBuildLocator/issues/86)