Simplify resolution of file edit to avoid unnecessary execute command (#868)
In the current implementation, handling the code action involves creating an execute command to edit the file. However, this leads to a deadlock in which the apply edit is waiting for the code action handling to complete and this can't complete until the edit action is applied.

It's unclear when this regression was introduced, but it's possible that recent changes in the synchronization model of the protocol are no longer allowing this approach.

This change is shortcutting this by returning an edit Microsoft.VisualStudio.LanguageServer.Protocol.CodeAction directly instead.