(Hacky) Port MDNode to LLVMSharp.
To do this, we introduce the expectation that there will only ever be a single instance of Context
which is true for QirGeneration in Q# today. This is needed because it's not possible to get the
context for a given metadata node using LLVMSharp / the LLVM C API. Instead, we assume the
ContextCache contains exactly 1 Context, and that all metadata was created with it.