Implement ToString for LlvmMetadata and MDString.
This works by first converting the metadata to a value via MetadataAsValue, which creates a value of type MetadataAsValue.
Then, we use LLVM C API LLVMPrintValueToString, passing this value. The implementation of this in LLVM checks if the value is
a MetadataAsValue, and if it is, it prints the metadata, only.

For MDString, we do the same thing, and LLVMGetMDString checks for
MetadataAsValue too.

https://llvm.org/doxygen/AsmWriter_8cpp_source.html#l04652