Add a GetOperandValue to MetadataOperandCollection.
This should allow us to read operands as value types, enabling some previously remove properties.

Previously, Ubiquity relied on always being able to convert a ValueAsMetadata to a Value, which does not seem possible with the LLVM C API.
Because the LLVM C API GetMDNodeOperands function returns Value refs, we should be able to just read the operands as values directly, rather
than relying on a conversion later.