Port Debug and Metadata API to LLVMBindings (#1026)
* Add Ubiquity DebugInfo and Metadata.

Note that this includes some changes to begin porting to LLVMSharp.

* Add DebugInfo builder to BitcodeModule.

* (Pass 1) Porting LLVM Debug and Metadata to LLVMSharp.

* (Pass 2) Porting LLVM Debug and Metadata to LLVMSharp.

* Port NamedMDNode to LLVMSharp.

* Remove LocationDescribes validation from DebugInfoBuilder.InsertValue.

Not supported by LLVM C API.

* Remove MDNode IsTemporary and IsResolved.

These are only used for validation (to throw when they aren't of the expected value), so there should be no functional change,
other than allowing clients to attempt invalid LLVM operations.

Reason: not supported by LLVM C API.

* Remove MDNode IsDistinct and IsUniqued.

Reason: not supported by LLVM C API.

* Remove DISubRange GetLowerBound.

Reason: not supported by LLVM C API.

* Remove DILocation InlinedAtScope.

Reason: not supported by LLVM C API.

* Remove DINode Tag.

This was checked only to return null when DIDerivedType operand properties are invoked on instances with the wrong tag.
I.e., this commit removed validation.

Reason: not supported by LLVM C API.

* Remove DebugInfoBuilder Finish for individual Subprograms.

Reason: not supported by LLVM C API.

* Remove DebugInfoBuilder ForwardDeclareFunction.

Reason: not supported by LLVM C API.

* Remove DIBasicType Encoding.

Reason: not supported by LLVM C API.

* Fix DebugType TryGetExtendedPropertyValue.

* Remove Subprogram Describes call.

Describes was removed earlier since it isn't supported by LLVM C API.

* (Hacky) Port MDNode to LLVMSharp.

To do this, we introduce the expectation that there will only ever be a single instance of Context
which is true for QirGeneration in Q# today. This is needed because it's not possible to get the
context for a given metadata node using LLVMSharp / the LLVM C API. Instead, we assume the
ContextCache contains exactly 1 Context, and that all metadata was created with it.

* Fix missing declaration in TypeRef.cs.

* Remove typo.

* Remove ValueAsMetadata Value property.

Removes properties on ValueAsMetadata nodes that were backed by this as well. Also removes an unused contructor from ModuleFlag
that used it as well. That constructor is marked internal and is also
unused in Llvm.NET.

* Implement ToString for LlvmMetadata and MDString.

This works by first converting the metadata to a value via MetadataAsValue, which creates a value of type MetadataAsValue.
Then, we use LLVM C API LLVMPrintValueToString, passing this value. The implementation of this in LLVM checks if the value is
a MetadataAsValue, and if it is, it prints the metadata, only.

For MDString, we do the same thing, and LLVMGetMDString checks for
MetadataAsValue too.

https://llvm.org/doxygen/AsmWriter_8cpp_source.html#l04652

* Remove old TODO comment.

* Implement LLVMValueRef.GetMDNodeOperands.

* Fix bug in IrFunction SetSubprogram.

An argument check was not correctly adjusted in an earlier commit.

* Fix MetadataOperandCollection GetOperand for null operands.

Previously, attempting to get an undefined operand would result in a crash. Now, we just return null if the operand isn't found.

* [Drop] For testing, implement Q# debug info for CU, Namespace, an functions.

Note that this isn't a complete implementation and is just to make sure the ported code is working. Namely, debug type info is not attempted and is just stubbed out for function sigs.

Metadata Operands are working, e.g. we can see proper values for various MDNode class properties (tested using C# immediate window).

* Add a GetOperandValue to MetadataOperandCollection.

This should allow us to read operands as value types, enabling some previously remove properties.

Previously, Ubiquity relied on always being able to convert a ValueAsMetadata to a Value, which does not seem possible with the LLVM C API.
Because the LLVM C API GetMDNodeOperands function returns Value refs, we should be able to just read the operands as values directly, rather
than relying on a conversion later.

* Remove old TODO.

* Make exception message correct for ContextCache.Single().

* Use implicit conversion to handle in LLVMMetadataRefExtensions.

* Remove empty string args from exception constructors.

* Improve error handling for operand access.

* Use ArgumentOutOfRangeException in MetadataOperandCollection.

Also add doc comment to GetOperandValue.

* Make ContextCache.Single() implementation thread safe.

Support for Single() can be turned on and off now easily with
preprocessor definition SINGLE_CONTEXT. To re-enable the original
support for multiple contexts, just undefine it. This will cause the
Single() method to disappear, as well as tracking of the single instance
within ContextCache, without additional code changes.

* Improve ContextCache documentation.

* Add doc comment summary tag for new LLVMSharpExtension methods.

* Add <see> tag for all LLVMSharpExtension doc comment mentions of underlying native LLVM methods.

* Add doc comments for BitcodeModule NamedMDNode methods.

* Revert "[Drop] For testing, implement Q# debug info for CU, Namespace, an functions."

This reverts commit deb781d078b32959060d719a773761d8d25871fc.

* Add back accidentally removed GlobalIndirectSymbol.IndirectSymbol setter.

* Add comment regarding single context requirement to all callsites.

* Remove Ubiquity TODO comments we don't plan to address.

* Wrap use of LLVMModuleFlagEntry in BitcodeModule in try...finally.

* [style] Simple integer types.

* [style] Fix spacing issues.

* [style] Simplify ref names.

* [style] Update file headers.

* [style] Fix global variable name casing.

* [style] Add Microsoft C header text.

* [style] Local vars start with lowercase.

* [style] Prefix local calls with 'this'.

* [style] Reorder usings.

* [style] Minor misc fixes.

* [style] Add trailing comma to enums.

* Add back accidentally removed CatchSwitch.ParentPad setter.

* Use thread local storage for LLVM context.

* Fix existing bug where GenerationContext is not managed by a using block.

Previously, GenerationContext.Dispose was not being called to clean up LLVM resources between compilation units.