Add support for QIR Generation of Library projects (#1062)
* Add support for QIR Generation of Library projects

This change adds support for generation of QIR from Library projects. When a Q# Library is compiled with `<QirGeneration>true</QirGeneration>` the compiler will make certain assumptions to enable output of usable QIR. First, it will treat every non-internal, non-generic callabe defined in the library as an effective entry point for the purpose of syntax tree trimming and monomorphization. Second, it will generate the QIR function for each specialization of these effective entry points without the `internal` linkage so that they can be invoked from outside of the compilation unit. And finally, it will NOT generate Interop or EntryPoint friendly wrappers for these effective entry points, to allow callers to instead invoke specific specializations directly while supporting any QIR-specific types, like `%Qubit*` or `%Tuple*`. This should enable new use cases of Q# libraries across FFI boundaries assuming the resulting library QIR gets compiled with the same runtime that it is eventually invoked with.

* Split IR files for resilient checks, run formatter

* Avoid propagating modified EntryPoints