Setting LSP document version in Language Server (#1085)
Each time that a document is change in an editor, the Language Server Protocol includes a document version number as part of its `textNotification/didChange` updates. This document version can be specified as part of WorkspaceEdit objects when setting up code actions.

Up to Visual Studio 16.9, it was optional by the LSP protocol to include a matching version, and the extension defaulted to a version of 1 in all cases. However, starting from 16.10, the protocol expects the version of the document to match the latest one provided by the LSP, otherwise code actions will be ignored.

This change keeps track of the last received version from the LSP and uses it when setting up code actions.