Remove call to initialize LLVM target (#1123)
The `GenerationContext` constructor included a call to Register the current plaform as the target for LLVM, which it turns out is unnecessary for generating target agnostic IR like we do. Further, this reduces the surface are of APIs that we utilize from LLVM, making it easy to identify the required libraries for QIR generation.