Create standalone llvm-build
This builds off of work from @idavis that establishes a custom LLVM pipeline build.
The output is then used along with ClangSharpPInvokeGenerator to create custom .NET
wrappers around the resulting binaries. The native binaries are then packaged so they
can be used as a dev-dependency and pulled into the LlvmBindings package later.