Create VSIX for VS Code extension using a version allowed by Marketplace (#1321)
Visual Studio Marketplace will not allow an extension with a version number that contains a release type suffix like '-beta'.

This blocks us from releasing directly a beta build as built. With this change, we allow the VSIX extension to be packaged using a version that only contains the numeric part of the version, while keeping the NuGet version of all components in full (including the release type suffix).