Remove `waitForInit` from LanguageServer
This change simplifies the synchronization of the LanguageServer by removing the `waitForInit` ManualResetEvent. Since LSP clients should not send further messages until they have received the response to init and the change adjusts the rpc listener to not connect until after the object is constructed, the synchronization should no longer be needed.