Always send telemetry events from Visual Studio and Language Server (#1493)
The QsCompiler and VisualStudioExtension solutions are being built with the TELEMETRY flag undefined. This is preventing telemetry events from being published. The flag should be set as a consequence of Assembly Constants containing SIGNED:TELEMETRY, which is set, but in those projects it's not.

For this release, we're directly enabling the telemetry code (on the release branch) while a proper solution is done in 'main' for the next release.