Add more utilties to LlvmBindings (#1530)
This change adds two utilities to the LlvmBindings for usability. First, it adds back the functionality for parsing `BitcodeModule` objects from a file via a static `LoadFrom` method. Second, it adds a `ReturnBlocks` convenience method to `IrFunction` objects to make it easier for a caller to request any blocks that are terminated by a return.